backend/requirements.txt
Django>=6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=2.0
django-cors-headers>=3.15
